Updated namespaces to separate TEAM from TEAM lib.
