+++ TEAM/JsonOutputHandling.cs
@@ -1,6 +1,8 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using DataWarehouseAutomation;
+using TEAM_Library;
+using Extension = DataWarehouseAutomation.Extension;
 
 namespace TEAM
 {
