+++ TEAM/LocalTeamEnvironmentConfiguration.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Text;
 using System.Windows.Forms;
+using TEAM_Library;
 
 namespace TEAM
 {
